feat: 애니메이션 위젯으로 정리
