+++ lib/domain/repository/diary/diary_repository.dart
@@ -2,7 +2,7 @@ import 'package:frontend/core/result.dart';
 import 'package:frontend/domain/model/diary/diary_data.dart';
 
 abstract class DiaryRepository {
-  Future<Result<bool>> saveDiary(DiaryData diary);
+  Future<Result<String>> saveDiary(DiaryData diary);
 
   Future<Result<bool>> updateDiary(DiaryData diary);
